% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_cut_data}
\alias{get_cut_data}
\alias{add.cutline}
\alias{wnom_adds_get_angles}
\alias{wnom_adds_get_cutlines}
\title{Build/extract cutline data}
\usage{
add.cutline(cutData, weight)

wnom_adds_get_angles(x, dims = c(1, 2), ...)

wnom_adds_get_cutlines(x, dims = c(1, 2), add_arrows = TRUE, rollcall_obj,
  ...)
}
\arguments{
\item{x}{A nomObj via wnominate::wnominate output}

\item{rollcall_obj}{An rollcall object from pscl package}
}
\value{
A data frame
}
\description{
Functions extract cutting line data for usage in 'tidier' frameworks.
}
