% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_wnominate_underlying}
\alias{get_wnominate_underlying}
\alias{add.cutline}
\alias{wnm_get_cutlines}
\alias{wnm_get_angles}
\title{Build/extract cutline data}
\usage{
add.cutline(cutData, weight)

wnm_get_cutlines(x, dims = c(1, 2), arrow_length, rollcall_obj, ...)

wnm_get_angles(x, dims = c(1, 2), ...)
}
\arguments{
\item{x}{A nomObj via wnominate::wnominate output}

\item{rollcall_obj}{A rollcall object from pscl package}
}
\value{
A data frame
}
\description{
Functions extract cutting line data for usage in 'tidier' frameworks.
}
