% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_voteview.R
\docType{data}
\name{scale_voteview}
\alias{scale_voteview}
\alias{vote}
\alias{scale_color_rollcall}
\alias{scale_color_party}
\alias{scale_shape_rollcall}
\title{NOMINATE/VoteView.com color palette}
\format{An object of class \code{factor} of length 6.}
\usage{
vote

scale_color_rollcall(...)

scale_color_party(...)

scale_shape_rollcall(...)
}
\description{
The standard color scheme for party affiliation and votes by party affiliation.
}
\keyword{datasets}
