% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnm_get_cutlines.R
\name{get_wnominate_underlying}
\alias{get_wnominate_underlying}
\alias{wnm_get_cutlines}
\title{Build/extract cutline data}
\usage{
wnm_get_cutlines(x, dims = c(1, 2), arrow_length = 0.05,
  vote_id = NULL, rollcall_obj, ...)
}
\arguments{
\item{x}{A nomObj via wnominate::wnominate output}

\item{rollcall_obj}{A rollcall object from pscl package}
}
\value{
A data frame
}
\description{
Functions extract cutting line data for usage in 'tidier' frameworks.
}
